[Common] - Elevated apps warning moved to common notifications - Warning added to AoT
