+++ src/common/notifications/dont_show_again.h
@@ -4,8 +4,8 @@
 
 namespace notifications
 {
-    const inline wchar_t CantDragElevatedDontShowAgainRegistryPath[] = LR"(SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\DontShowMeThisDialogAgain\{e16ea82f-6d94-4f30-bb02-d6d911588afd})";
-    const inline int64_t CantDragElevatedDisableIntervalInDays = 30;
+    const inline wchar_t ElevatedDontShowAgainRegistryPath[] = LR"(SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\DontShowMeThisDialogAgain\{e16ea82f-6d94-4f30-bb02-d6d911588afd})";
+    const inline int64_t ElevatedDisableIntervalInDays = 30;
 
     const inline wchar_t PreviewModulesDontShowAgainRegistryPath[] = LR"(SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\DontShowMeThisDialogAgain\{7e29e2b2-b31c-4dcd-b7b0-79c078b02430})";
     const inline int64_t PreviewModulesDisableIntervalInDays = 30;
