+++ src/common/notifications/resource.base.h
@@ -1,0 +1,9 @@
+//{{NO_DEPENDENCIES}}
+// Microsoft Visual C++ generated include file.
+// Used by notifications.rc
+
+//////////////////////////////
+// Non-localizable
+
+// Non-localizable
+//////////////////////////////
