+++ src/modules/fancyzones/FancyZonesLib/Resources.resx
@@ -217,16 +217,6 @@
     <value>FancyZones</value>
     <comment>FancyZones is a product name, keep as is.</comment>
   </data>
-  <data name="Cant_Drag_Elevated" xml:space="preserve">
-    <value>We've detected an application running with administrator privileges. This will prevent certain interactions with these applications.</value>
-    <comment>administrator is context of user account.</comment>
-  </data>
-  <data name="Cant_Drag_Elevated_Learn_More" xml:space="preserve">
-    <value>Learn more</value>
-  </data>
-  <data name="Cant_Drag_Elevated_Dialog_Dont_Show_Again" xml:space="preserve">
-    <value>Don't show again</value>
-  </data>
   <data name="Keyboard_Listener_Error" xml:space="preserve">
     <value>Failed to install keyboard listener.</value>
   </data>
